Refactoring to package `transformers.onnx`
