+++ src/transformers/models/gpt2/configuration_gpt2.py
@@ -15,7 +15,8 @@
 # limitations under the License.
 """ OpenAI GPT-2 configuration """
 
-from ...configuration_utils import PretrainedConfig, OnnxConfig, OnnxVariable
+from ...configuration_utils import PretrainedConfig
+from ...onnx import OnnxConfig, OnnxVariable
 from ...utils import logging
