+++ src/transformers/models/longformer/__init__.py
@@ -22,7 +22,7 @@ from ...file_utils import _LazyModule, is_tf_available, is_tokenizers_available,
 
 
 _import_structure = {
-    "configuration_longformer": ["LONGFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP", "LongformerConfig"],
+    "configuration_longformer": ["LONGFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP", "LONGFORMER_ONNX_CONFIG", "LongformerConfig"],
     "tokenization_longformer": ["LongformerTokenizer"],
 }
 
@@ -57,7 +57,7 @@ if is_tf_available():
 
 
 if TYPE_CHECKING:
-    from .configuration_longformer import LONGFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP, LongformerConfig
+    from .configuration_longformer import LONGFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP, LONGFORMER_ONNX_CONFIG, LongformerConfig
     from .tokenization_longformer import LongformerTokenizer
 
     if is_tokenizers_available():
