+++ src/transformers/models/longformer/configuration_longformer.py
@@ -16,6 +16,7 @@
 
 from typing import List, Union
 
+from ...onnx import OnnxConfig, OnnxVariable
 from ...utils import logging
 from ..roberta.configuration_roberta import RobertaConfig
 
@@ -69,3 +70,33 @@ class LongformerConfig(RobertaConfig):
     def __init__(self, attention_window: Union[List[int], int] = 512, sep_token_id: int = 2, **kwargs):
         super().__init__(sep_token_id=sep_token_id, **kwargs)
         self.attention_window = attention_window
+
+
+LONGFORMER_ONNX_CONFIG = OnnxConfig(
+    inputs=[
+        OnnxVariable("input_ids", {0: "batch", 1: "sequence"}, repeated=1, value=None),
+        OnnxVariable("attention_mask", {0: "batch", 1: "sequence"}, repeated=1, value=None),
+    ],
+    outputs=[
+        OnnxVariable("last_hidden_state", {0: "batch", 1: "sequence"}, repeated=1, value=None),
+        OnnxVariable("pooler_output", {0: "batch"}, repeated=1, value=None),
+    ],
+    runtime_config_overrides=None,
+    use_external_data_format=False,
+    minimum_required_onnx_opset=12,
+    optimizer="bert",
+    optimizer_features={
+        "enable_gelu": True,
+        "enable_layer_norm": True,
+        "enable_attention": True,
+        "enable_skip_layer_norm": True,
+        "enable_embed_layer_norm": True,
+        "enable_bias_skip_layer_norm": True,
+        "enable_bias_gelu": True,
+        "enable_gelu_approximation": False,
+    },
+    optimizer_additional_args={
+        "num_heads": "$config.num_attention_heads",
+        "hidden_size": "$config.hidden_size"
+    }
+)
