+++ src/transformers/models/roberta/configuration_roberta.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """ RoBERTa configuration """
-from ... import OnnxConfig, OnnxVariable
+from ...onnx import OnnxConfig, OnnxVariable
 from ...utils import logging
 from ..bert.configuration_bert import BertConfig
